+++ Project/src/main/java/DomainLayer/Shop/Discount/PolicyLeaf.java
@@ -6,14 +6,14 @@ import DomainLayer.Item.ItemCategory;
 
 public class PolicyLeaf implements Policy{
 
-    private final TriPredicate<Map<Integer,Integer>,Map<Integer,Integer>,Map<Integer,ItemCategory>> predicate;
+    private final TriPredicate<Map<Integer,Integer>,Map<Integer,Double>,Map<Integer,ItemCategory>> predicate;
 
-    public PolicyLeaf(TriPredicate<Map<Integer,Integer>, Map<Integer,Integer>,Map<Integer,ItemCategory>> predicate) {
+    public PolicyLeaf(TriPredicate<Map<Integer,Integer>, Map<Integer,Double>,Map<Integer,ItemCategory>> predicate) {
         this.predicate = predicate;
     }
 
     @Override
-    public boolean test(Map<Integer,Integer> items, Map<Integer,Integer> prices, Map<Integer,ItemCategory> itemsCategory) {
+    public boolean test(Map<Integer,Integer> items, Map<Integer,Double> prices, Map<Integer,ItemCategory> itemsCategory) {
         return predicate.test(items, prices, itemsCategory);
     }
