+++ Project/src/main/java/DomainLayer/Shop/Discount/SingleDiscount.java
@@ -30,7 +30,7 @@ public class SingleDiscount implements Discount {
 
 
     @Override
-    public Map<Integer, Integer> applyDiscounts(Map<Integer, Integer> items, Map<Integer, AtomicInteger> prices, Map<Integer, Integer> itemsDiscountedPrices, Map<Integer, ItemCategory> itemsCategory) {
+    public Map<Integer, Double> applyDiscounts(Map<Integer, Integer> items, Map<Integer, AtomicInteger> prices, Map<Integer, Double> itemsDiscountedPrices, Map<Integer, ItemCategory> itemsCategory) {
         // item-specific: calculate discount only on that item
         Integer qty = items.get(itemId);
         if (qty == null || qty <= 0) {
@@ -43,12 +43,12 @@ public class SingleDiscount implements Discount {
         }
 
         if(!isDouble){
-            int itemPrice = prices.get(itemId).get();
-            int discountedPrice = itemPrice * (100 - percentage) / 100;
+            double itemPrice = prices.get(itemId).get();
+            double discountedPrice = itemPrice * (100 - percentage) / 100;
             itemsDiscountedPrices.put(itemId,Math.min(itemsDiscountedPrices.get(itemId), discountedPrice));
         }else{
-            int itemPrice = itemsDiscountedPrices.get(itemId);
-            int discountedPrice = itemPrice * (100 - percentage) / 100;
+            double itemPrice = itemsDiscountedPrices.get(itemId);
+            double discountedPrice = itemPrice * (100 - percentage) / 100;
             itemsDiscountedPrices.put(itemId, discountedPrice);
         }
           
@@ -56,7 +56,7 @@ public class SingleDiscount implements Discount {
     }
 
     @Override
-    public boolean checkPolicies(Map<Integer,Integer> items, Map<Integer,Integer> prices, Map<Integer,ItemCategory> itemsCategory) {
+    public boolean checkPolicies(Map<Integer,Integer> items, Map<Integer,Double> prices, Map<Integer,ItemCategory> itemsCategory) {
         if(policyHead == null) {
             return true; // No policies to check
         }
